// insurancehub.client/src/app/shared/models/custom-labels-config.dto.ts
export class CustomLabelsConfig {
  Tax: string = '';
  HS_Code: string = '';
  PAN_No: string = '';
}

export class BillingReportCustomLabel {
  PatientCategory: string = '';
  Scheme: string = '';
}

export class AdmissionReportCustomLabel {
  PatientCategory: string = '';
  Scheme: string = '';
}